Fix noise pixels at display edge
This just nudges the DisplayOffset two pixels to make the noise
rows on the left of the display go away (at least on my hardware,
AdaFruit 64x128 OLED FeatherWing). I have no idea why this is
needed, since the C++ driver which worked on the same hardware
appears to use 0x60.

This is not tested on any other hardware, so I'm not planning to
make it public. But if you ended up here to find a solution to
this specific problem on this specific hardware, have at it.